backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
mysql-connector-python
PyPDF2
rapidfuzz
requests
werkzeug